+++ src/bika/lims/browser/batch/analysisrequests.py
@@ -15,18 +15,13 @@
 # this program; if not, write to the Free Software Foundation, Inc., 51
 # Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #
-# Copyright 2018-2024 by it's authors.
+# Copyright 2018-2025 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-from bika.lims import api
-from bika.lims.browser.analysisrequest import AnalysisRequestsView as BaseView
+# flake8: noqa
 
+from senaite.core.browser.batches.samples import \
+    SamplesView as AnalysisRequestsView
+from zope import deprecation
 
-class AnalysisRequestsView(BaseView):
-
-    def __init__(self, context, request):
-        super(AnalysisRequestsView, self).__init__(context, request)
-        self.contentFilter = {"portal_type": "AnalysisRequest",
-                              "getBatchUID": api.get_uid(self.context),
-                              "sort_on": "created",
-                              "sort_order": "reverse"}
+deprecation.deprecated("AnalysisRequestsView", "Moved to senaite.core.browser.batches.samples.SamplesView")