+++ src/bika/lims/content/analysisrequest.py
@@ -15,7 +15,7 @@
 # this program; if not, write to the Free Software Foundation, Inc., 51
 # Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #
-# Copyright 2018-2024 by it's authors.
+# Copyright 2018-2025 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
 import base64
@@ -1516,7 +1516,10 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
 
         # Don't add analyses from profiles during sample creation.
         # In this case the required analyses are added afterwards explicitly.
-        if not api.is_temporary(self):
+        #
+        # Also only add analyses if a profile (value) is selected:
+        # https://github.com/senaite/senaite.core/pull/2672
+        if value and not api.is_temporary(self):
             # get the profiles
             profiles = map(api.get_object_by_uid, uids)
             # get the current set of analyses/services