+++ webpack/app/widgets/queryselect/widget.js
@@ -119,18 +119,53 @@ class QuerySelectWidgetController extends React.Component {
 
   /*
    * Throttle wrapper
+   *
+   * @returns {Promise}
    */
   debounce(func) {
-    let timer;
+    let timer = null;
+    const context = this;
+    let deferred = null;
+
     return function (...args) {
-      const context = this;
-      if (timer) clearTimeout(timer);
+      // clear the timer
+      if (timer) {
+        clearTimeout(timer);
+      }
+
+      // deferred promise that is returned to the caller
+      deferred = this.create_deferred()
+
       timer = setTimeout(() => {
+        // resolve the promise
+        Promise.resolve(func.apply(context, args))
+               .then(deferred.resolve)
+               .catch(deferred.reject)
+        // reset timer and promise
         timer = null;
-        func.apply(context, args);
+        deferred = null;
       }, 500);
+
+      return deferred.promise;
     };
   };
+
+
+  /*
+   * Helper method to create a deferred object
+   *
+   * @returns {Object} of promise, resolve, reject
+   */
+  create_deferred() {
+    let resolve;
+    let reject;
+    const promise = new Promise((res, rej) => {
+      resolve = res;
+      reject = rej;
+    });
+    return { promise, resolve, reject };
+  }
+
 
   /*
    * Fix overflow at the bottom or at the right of the container
@@ -242,6 +277,26 @@ class QuerySelectWidgetController extends React.Component {
 
 
   /*
+   * Check if the passed object is an array
+   *
+   * @returns {Boolean} true/false if object is an Array
+   */
+  is_array(obj) {
+    return Object.prototype.toString.call(obj) === '[object Array]';
+  }
+
+
+  /*
+   * Check if the passed object is an object
+   *
+   * @returns {Boolean} true/false if object is an Object
+   */
+  is_object(obj) {
+    return Object.prototype.toString.call(obj) === '[object Object]'
+  }
+
+
+  /*
    * Returns if the field accepts single values only
    *
    * @returns {Boolean} true/false if  values are allowed
@@ -262,6 +317,16 @@ class QuerySelectWidgetController extends React.Component {
 
 
   /*
+   * Check if the field has a value set
+   *
+   * @returns {Boolean} true/false if a value is set or not
+   */
+  has_value() {
+    return this.state.values.length > 0;
+  }
+
+
+  /*
    * Checks if the field should be rendered as disabled
    *
    * @returns {Boolean} true/false if the widget is disabled
@@ -273,7 +338,7 @@ class QuerySelectWidgetController extends React.Component {
     if (this.state.readonly) {
       return true;
     }
-    if (this.is_single_valued() && this.state.values.length > 0) {
+    if (this.is_single_valued() && this.has_value()) {
       return true;
     }
     return false;
@@ -292,7 +357,7 @@ class QuerySelectWidgetController extends React.Component {
     if (this.state.readonly) {
       return false;
     }
-    if (this.is_single_valued() && this.state.values.length > 0) {
+    if (this.is_single_valued() && this.has_value()) {
       return false;
     }
     return true;
@@ -316,6 +381,16 @@ class QuerySelectWidgetController extends React.Component {
    */
   get_query() {
     return this.state.query;
+  }
+
+
+  /*
+   * Returns the item value key
+   *
+   * @returns {String} name
+   */
+  get_item_value_key() {
+    return this.state.value_key;
   }
 
 
@@ -515,6 +590,18 @@ class QuerySelectWidgetController extends React.Component {
       return;
     }
     console.debug("QuerySelectWidgetController::select:value:", value);
+    // clear any previous result for single value fields
+    if (this.is_single_valued() && this.has_value()) {
+      this.state.values = [];
+    }
+    // handle object values
+    if (this.is_object(value)) {
+      return this.select(value[this.get_item_value_key()]);
+    }
+    // handle array values
+    if (this.is_array(value)) {
+      return value.forEach((item) => { this.select(item) })
+    }
     // create a copy of the selected values
     let values = [].concat(this.state.values);
     // Add the new value if it is not selected yet
@@ -556,12 +643,12 @@ class QuerySelectWidgetController extends React.Component {
    * Add/remove the focused result
    *
    */
-  select_focused(searchvalue) {
+  select_focused() {
     console.debug("QuerySelectWidgetController::select_focused");
     let focused = this.state.focused;
     let result = this.state.results.at(focused);
     if (result) {
-      let value = result[this.state.value_key];
+      let value = result[this.get_item_value_key()];
       if (this.state.values.indexOf(value) == -1) {
         this.select(value);
       } else {
@@ -743,7 +830,7 @@ class QuerySelectWidgetController extends React.Component {
     //       for stored UIDs when the edit form is initially rendered
     let records = Object.assign(this.state.records, {})
     for (let item of items) {
-      let value = item[this.state.value_key];
+      let value = item[this.get_item_value_key()];
       records[value] = item;
     }
 
@@ -797,7 +884,7 @@ class QuerySelectWidgetController extends React.Component {
     promise.then((data) => {
       let items = data.items || [];
       for (let item of items) {
-        let value = item[this.state.value_key];
+        let value = item[this.get_item_value_key()];
         records[value] = item;
       }
       this.toggle_loading(false);
@@ -857,7 +944,7 @@ class QuerySelectWidgetController extends React.Component {
    */
   on_sync(event) {
     let values = event.detail.values || [];
-    let is_array = {}.toString.call( values ) === '[object Array]';
+    let is_array = this.is_array(values)
     if (!is_array) {
       values = values.split("/n");
     }
@@ -894,7 +981,7 @@ class QuerySelectWidgetController extends React.Component {
             className="queryselectwidget-results-container position-fixed shadow-lg border border-light rounded-lg bg-white mt-2 p-1"
             columns={this.get_columns()}
             values={this.state.values}
-            value_key={this.state.value_key}
+            value_key={this.get_item_value_key()}
             searchterm={this.state.searchterm}
             width={this.state.results_table_width}
             results={this.state.results}